pyproject.toml
[project]
name = "notion-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.31.0",
    "cryptography>=46.0.5",
    "fastapi>=0.135.1",
    "fastmcp>=3.1.1",
    "google-cloud-documentai>=3.11.0",
    "httpx>=0.28.1",
    "mcp>=1.2.0",
    "pillow>=12.1.1",
    "pydantic-settings>=2.13.1",
    "pymupdf>=1.27.2.2",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.2.2",
    "python-jose>=3.5.0",
    "supabase>=2.28.2",
    "uvicorn>=0.42.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
